Experimento2/3004-Monitor/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.0
packaging==23.2
psycopg2-binary==2.9.7
python-dotenv==1.0.1
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
typing_extensions==4.10.0
Werkzeug==3.0.1
